+++ CHANGELOG.md
@@ -4,7 +4,8 @@ Changes to Calva.
 
 ## [Unreleased]
 
-- Fix: [set calva:cursorInWhitespaceAfterComment if the cursor is in whitespace after a comment's newline](https://github.com/BetterThanTomorrow/calva/issues/2917)
+- [Make delete word backwards more precise and intuitive when the cursor is in the whitespace after a line comment](https://github.com/BetterThanTomorrow/calva/issues/2917)
+- [remove duplicate paredit definition in Backward to List Start keybinding](https://github.com/BetterThanTomorrow/calva/issues/2916)
 
 ## [2.0.525] - 2025-08-30